classification_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.0.1
mlflow==2.20.2
python-dotenv==1.0.1
python-multipart==0.0.20